// src/content/types.tsx
export type ICardSkill = {
  title: string;
  content: string;
  expandedContent: string;
};
export type ICardContent = {
  css: ICardSkill;
  html: ICardSkill;
  git: ICardSkill;
  figma: ICardSkill;
  jquery: ICardSkill;
  cpp: ICardSkill;
  react: ICardSkill;
  typescript: ICardSkill;
  gsap: ICardSkill;
  javascript: ICardSkill;
  nodejs: ICardSkill;
  jss: ICardSkill;
};